# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Notifications API
  version: 1.0.0
  description: |
    HTTP surface for the notifications service shipped through Phase 4 per
    `ARCHITECTURE_v3.md` and the `Software Engineer Assessment - Golang.md`
    brief. Accepts notifications (single or batch), returns notification
    state, supports best-effort cancellation, and lists / filters
    notifications. Notifications dispatch across SMS, email, and push
    channels via a Postgres-backed outbox + per-channel Kafka topics +
    per-channel worker pools.

    All endpoints are unauthenticated by design (out of scope per
    `docs/design/03-api.md` §Conventions). Idempotency is enforced via
    the request-body `idempotency_key` (UUIDv4 canonical lowercase) on
    every notification.
  contact:
    name: Notifications API maintainers
    url: https://github.com/tarkandikmen/notifications
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development via docker-compose
...
paths:
...
  /v1/notifications/{id}/cancel:
    post:
      operationId: cancelNotification
      summary: Best-effort cancel of a notification
      description: |
        Runs the cancel transition `T3` (PENDING → CANCELLED, emits
        `events.notification`) or `T11` (DISPATCHED → CANCELLED, no
        emit) per `docs/design/02-state-machine.md` §Transitions.

        Already-CANCELLED rows return 200 with the same body shape
        (idempotent no-op). Hard-terminal rows (DELIVERED / FAILED)
        return 409 `terminal_state` with the observed status in
        `details[0].current_status`.

        Best-effort semantics: a cancel applied to a DISPATCHED row may
        be silently overwritten by a worker outcome (T4–T8) on the
        row's current `attempt`. The realized state surfaces via a
        subsequent GET; the cancel response itself reflects the row at
        cancel-commit time. See `docs/design/02-state-machine.md`
        §Cancellation race for the full contract.
      tags: [Notifications]
      parameters:
        - $ref: '#/components/parameters/NotificationId'
      responses:
        '200':
          $ref: '#/components/responses/OK200Notification'
        '404':
          $ref: '#/components/responses/NotFound404'
        '409':
          $ref: '#/components/responses/Conflict409Terminal'
        '500':
          $ref: '#/components/responses/InternalServerError500'
  /v1/batches/{id}:
    get:
      operationId: getBatch
      summary: Fetch every notification in a batch
      description: |
        Returns every notification sharing the supplied `batch_id`,
        ordered by `id ASC` (UUIDv7 ids are time-ordered, so this
        matches the original request order). Capped at `batch_max=1000`
        rows by construction; no pagination.

        Returns 404 when no row matches the id (whether unknown or
        malformed). Items use the no-attempts notification
        representation; clients that need delivery attempts must
        follow up per row.
      tags: [Notifications]
      parameters:
        - $ref: '#/components/parameters/BatchId'
      responses:
        '200':
          $ref: '#/components/responses/OK200Batch'
        '404':
          $ref: '#/components/responses/NotFound404'
        '500':
          $ref: '#/components/responses/InternalServerError500'
components:
  parameters:
    NotificationId:
      name: id
      in: path
      required: true
      description: Notification UUIDv7.
      schema:
        type: string
        format: uuid
    BatchId:
      name: id
      in: path
      required: true
      description: Batch UUIDv7.
      schema:
        type: string
        format: uuid
  schemas:
...
    Channel:
      type: string
      enum: [sms, email, push]
      description: Delivery channel. Drives per-channel topic + worker pool routing.
    Status:
      type: string
      enum: [PENDING, DISPATCHED, DELIVERED, FAILED, CANCELLED]
      description: |
        Notification lifecycle status per
        `docs/design/02-state-machine.md` §States. `DELIVERED` and
        `FAILED` are hard-terminal; `CANCELLED` is soft-terminal (a
        worker may overwrite a `DISPATCHED → CANCELLED` cancel during
        the cancellation race).
    Priority:
      type: string
      enum: [low, normal, high]
      description: |
        Priority bucket. Stored as int (`low=0`, `normal=1`, `high=2`)
        per `docs/design/01-schema.md` §1; the API accepts and returns
        the string form.
    FailureReason:
      type: string
      enum: [max_attempts_exceeded, permanent_error, unprocessable_message]
      description: |
        Set when `status=FAILED`. `max_attempts_exceeded` covers T7
        (worker exhausted retries) and T10 (reaper terminal-fail);
        `permanent_error` covers T6; `unprocessable_message` covers T8.
...
    NotificationRepresentation:
      type: object
      description: |
        The no-attempts notification shape returned by list, batch-get,
        and cancel responses per `docs/design/03-api.md` §Notification
        representation.
      required:
        - id
        - channel
        - recipient
        - priority
        - status
        - attempt
        - eligible_at
        - idempotency_key
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Notification UUIDv7.
        batch_id:
          type: string
          format: uuid
          description: Omitted when null (single-create rows).
        channel:
          $ref: '#/components/schemas/Channel'
        recipient:
          type: string
        priority:
          $ref: '#/components/schemas/Priority'
        content:
          type: string
          description: Omitted when null. Exactly one of `content` / `template` is non-null per row.
        template:
          type: string
          description: Omitted when null. Phase 6.
        template_data:
          type: object
          additionalProperties: true
          description: Omitted when null. Phase 6.
        status:
          $ref: '#/components/schemas/Status'
        attempt:
          type: integer
          minimum: 0
          description: |
            Number of dispatch claims so far. Incremented by the
            dispatcher in T2 only; every other transition leaves it
            unchanged per `docs/design/02-state-machine.md`
            §Counter discipline.
        eligible_at:
          type: string
          format: date-time
          description: |
            The dispatcher's claim-eligibility column. Equals
            `scheduled_at` when the request was scheduled; otherwise
            equals `created_at` initially.
        scheduled_at:
          type: string
          format: date-time
          description: Omitted when null. Forensic copy of the request's `scheduled_at`.
        failure_reason:
          allOf:
            - $ref: '#/components/schemas/FailureReason'
          description: Omitted when null. Set when `status=FAILED`.
        idempotency_key:
          type: string
          format: uuid
          description: Client-supplied UUIDv4 (canonical lowercase).
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          description: |
            Maintained by the `notifications_set_updated_at` trigger
            on every UPDATE.
...
    BatchGetResponse:
      type: object
      required: [batch_id, notifications]
      properties:
        batch_id:
          type: string
          format: uuid
          description: Echo of the path-supplied batch id.
        notifications:
          type: array
          description: |
            Every notification sharing this `batch_id`, ordered by
            `id ASC`. Capped at `batch_max=1000` by construction.
          items:
            $ref: '#/components/schemas/NotificationRepresentation'
    ErrorEnvelope:
      type: object
      required: [error]
      description: Canonical wrapper for every non-2xx response.
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
    ErrorBody:
      type: object
      required: [code, message]
      description: |
        Inner payload of `ErrorEnvelope`. `details` is omitted (404,
        413, 500) or carries a single concrete entry shape per code:
        `validation_failed` → `FieldIssue[]`,
        `duplicate_idempotency_keys` → `IdempotencyConflictDetail[]`,
        `terminal_state` → `[TerminalStateDetail]`.
      properties:
        code:
          type: string
          enum:
            - validation_failed
            - duplicate_idempotency_keys
            - terminal_state
            - not_found
            - payload_too_large
            - internal_error
        message:
          type: string
          description: Human-readable message; not stable across versions.
        details:
          type: array
          description: |
            Polymorphic per `code`. Encoded as a heterogeneous array
            so each error code can supply its own entry shape.
          items:
            oneOf:
              - $ref: '#/components/schemas/FieldIssue'
              - $ref: '#/components/schemas/IdempotencyConflictDetail'
              - $ref: '#/components/schemas/TerminalStateDetail'
    FieldIssue:
      type: object
      required: [path, issue]
      description: |
        One entry in the `validation_failed` `details[]` array. `path`
        names the offending request field (e.g., `channel`,
        `notifications[3].idempotency_key`); `issue` describes the
        rule that failed.
      properties:
        path:
          type: string
          example: notifications[0].channel
        issue:
          type: string
          example: required
    IdempotencyConflictDetail:
      type: object
      required: [idempotency_key, existing_id, status]
      description: |
        One entry in the `duplicate_idempotency_keys` `details[]`
        array. Carries the existing row's id and status so the client
        can reconcile without an extra GET.
      properties:
        idempotency_key:
          type: string
          format: uuid
          description: The conflicting key (canonical lowercase UUIDv4).
        existing_id:
          type: string
          format: uuid
          description: The existing row's UUIDv7.
        status:
          $ref: '#/components/schemas/Status'
    TerminalStateDetail:
      type: object
      required: [current_status]
      description: |
        Single-element entry in the `terminal_state` `details[]`
        array. Surfaced when a cancel hits a hard-terminal row.
      properties:
        current_status:
          type: string
          enum: [DELIVERED, FAILED]
          description: |
            The observed terminal status. `CANCELLED` does not appear
            here — already-cancelled rows return 200 (idempotent
            no-op) per `docs/design/03-api.md`
            §POST /v1/notifications/{id}/cancel.
  responses:
...
    OK200Notification:
      description: |
        Notification representation without nested delivery attempts.
        Used by cancel responses per
        `docs/design/03-api.md` §Notification representation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotificationRepresentation'
...
    OK200Batch:
      description: All notifications belonging to the supplied batch id.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BatchGetResponse'
...
    NotFound404:
      description: |
        No matching resource. `error.code` is `not_found`; `details`
        is omitted. Returned for unknown ids and for malformed
        (non-UUID) ids.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: not_found
              message: notification not found
...
    Conflict409Terminal:
      description: |
        Cancel target is in a hard-terminal state (`DELIVERED` or
        `FAILED`). `error.code` is `terminal_state`;
        `error.details[]` carries a single `TerminalStateDetail`
        with the observed status.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: terminal_state
              message: notification is in a terminal state and cannot be cancelled
              details:
                - current_status: DELIVERED
...
    InternalServerError500:
      description: |
        Unexpected server-side failure. `error.code` is
        `internal_error`; `details` is omitted. The error is logged
        with structured context for operator investigation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: internal_error
              message: internal error